requirements.txt
whisper==1.1.10
faster_whisper==0.5.1
pyannote-audio
unstructured==0.7.1
pdf2image==1.16.3
wordcloud==1.9.1.1
imageio==2.28.1
ffmpeg==1.4
accelerate
modelscope
langchain==0.0.174
gradio==3.20.1
numpy
tqdm
more-itertools
transformers>=4.19.0
ffmpeg-python==0.2.0
streamlink
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1
zhipuai==2.1.4.20230812
python-dotenv